Remove forcely instead of waiting for the socket file to be deleted in test (#158)
The socket file will be deleted after server process is finished in Socket#unix_server_loop, but sometimes it is not deleted and remain. Then, test is failed such as https://github.com/ruby/vscode-rdbg/actions/runs/4350570271/jobs/7601404118\#step:9:80. That's why we need to remove forcibly if socket file exists.