Wait for clearing the console of all text
ElementClickInterceptedError is sometimes occurred in test. This error occurred when "context-view-block" element is clicked. "context-view-block" element seems to be created to prevent users from clicking on a view in the middle of rendering. One possible cause of ElementClickInterceptedError is that the view in the middle of rendering is clicked after clearing the console of all text.